--- classes/enemy.py
@@ -1,50 +1,0 @@
-# Standard Library Imports
-import random
-
-# Third Party Imports
-import pygame
-
-# Local Application Imports
-from classes.baseobject import BaseObject
-from configurations import DEFAULT_ENEMY_DROP, DEFAULT_ENEMY_SPEED
-
-
-class Enemy(BaseObject):
-
-    IMG = pygame.transform.scale(
-        pygame.image.load("./images/alien.png"), (56, 56)
-    )
-    IMG_WIDTH = IMG.get_rect().size[0]
-
-    def __init__(self, screen_boundary_x):
-        super().__init__(
-            x_cord=random.randint(0, screen_boundary_x - self.IMG_WIDTH),
-            y_cord=random.randint(50, 100),
-        )
-        self.x_cord_change = DEFAULT_ENEMY_SPEED
-        self.y_cord_change = DEFAULT_ENEMY_DROP
-
-    @classmethod
-    def from_manual_cords(
-        cls,
-        x_cord,
-        y_cord,
-        x_cord_change=DEFAULT_ENEMY_SPEED,
-        y_cord_change=DEFAULT_ENEMY_DROP,
-    ):
-        """Generate an enemy at given coordinates with change.
-
-        Examples
-        ------
-        >>> enemies = [Enemy.from_manual_cords(500, 470, 0.2, 0)]
-
-        """
-        enemy = cls(cls.IMG_WIDTH)
-        enemy.x_cord = x_cord
-        enemy.y_cord = y_cord
-        if(x_cord_change is not DEFAULT_ENEMY_SPEED):
-            enemy.x_cord_change = x_cord_change
-        if(y_cord_change is not DEFAULT_ENEMY_DROP):
-            enemy.y_cord_change = y_cord_change
-        return enemy
-
